Restructured `utils.py` into a testing `/utils` directory.
