+++ tests/utils/__init__.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+
+#
+# Unless explicitly stated otherwise all files in this repository are licensed
+# under the Apache 2 License.
+#
+# This product includes software developed at Datadog
+# (https://www.datadoghq.com/).
+#
+# Copyright 2018 Datadog, Inc.
+#
+
+import uuid
+
+default_board_id = str(uuid.uuid1())
+default_list_id = str(uuid.uuid1())
+default_repo_id = 100001
+
+from tests.utils.general import *
+from tests.utils.database import *
